moved first user message handling to service
