+++ src/js/services/auth-service.js
@@ -4,8 +4,5 @@ import { core } from 'services/core';
 import { setUser } from 'actions/user-actions';
 
 export function login(props) {
-    return core().appUsers.init(props).catch((e) => {
-        console.log(e); //eslint-disable-line no-console
-        throw e;
-    });
+    return core().appUsers.init(props);
 }
