+++ src/js/services/conversation-service.js
@@ -1,16 +1,30 @@
 import { store } from 'stores/app-store';
 import { addMessage, setConversation } from 'actions/conversation-actions';
-import { updateReadTimestamp as updateReadTimestampAction } from 'actions/app-state-actions';
+import { updateReadTimestamp as updateReadTimestampAction, showSettingsNotification } from 'actions/app-state-actions';
 import { setFayeSubscription, unsetFayeSubscription } from 'actions/faye-actions';
 import { core } from 'services/core';
 import { immediateUpdate } from 'services/user-service';
 import { initFaye } from 'utils/faye';
 import { storage } from 'utils/storage';
 
+export function handleFirstUserMessage(response) {
+    let state = store.getState();
+    if (state.appState.settingsEnabled && !state.user.email) {
+        let appUserMessageCount = state.conversation.messages.filter(message => message.role === 'appUser').length;
+
+        if (appUserMessageCount === 1) {
+            // should only be one message from the app user
+            store.dispatch(showSettingsNotification());
+        }
+    }
+
+    return response;
+}
+
 export function sendMessage(text) {
     var sendFn = () => {
-        // add an id just to be able to reconcile the message
-        // with the server response
+        // add an id just to please React
+        // this message will be replaced by the real one on the server response
         const message = {
             _id: Math.random(),
             text: text,
@@ -24,10 +38,7 @@ export function sendMessage(text) {
         return core().conversations.sendMessage(user._id, message).then((response) => {
             store.dispatch(setConversation(response.conversation));
             return response;
-        }).catch((e) => {
-            console.log(e); //eslint-disable-line no-console
-            throw e;
-        });
+        }).then(handleFirstUserMessage);
     };
 
     var promise = immediateUpdate(store.getState().user);
@@ -61,9 +72,6 @@ export function getConversation() {
 
         store.dispatch(setConversation(response.conversation));
         return response;
-    }).catch((e) => {
-        console.log(e); //eslint-disable-line no-console
-        throw e;
     });
 }
