+++ src/js/services/user-service.js
@@ -30,9 +30,6 @@ export function immediateUpdate(props) {
     return isDirty ? core().appUsers.update(user._id, props).then((response) => {
         store.dispatch(setUser(response.appUser));
         return response;
-    }).catch((e) => {
-        console.error(e); //eslint-disable-line no-console
-        throw e;
     }) : Promise.resolve({
         user: user
     });
@@ -68,8 +65,5 @@ export function trackEvent(eventName, userProps) {
         }
 
         return response;
-    }).catch((e) => {
-        console.log(e); //eslint-disable-line no-console
-        throw e;
     });
 }
