+++ src/js/stores/configure-prod.js
@@ -1,9 +1,6 @@
-import { createStore, applyMiddleware } from 'redux';
+import { createStore } from 'redux';
 import { RootReducer } from 'reducers/root-reducer';
-import { firstMessage } from 'stores/middlewares/messages';
-
-const finalCreateStore = applyMiddleware(firstMessage)(createStore);
 
 export function configureStore(initialState) {
-    return finalCreateStore(RootReducer, initialState);
+    return createStore(RootReducer, initialState);
 }
