--- src/js/utils/redux.js
@@ -1,6 +1,0 @@
-import createLogger from 'redux-logger';
-import { createStore, applyMiddleware } from 'redux';
-
-export const createStoreWithMiddleware = applyMiddleware(
-  createLogger()
-)(createStore);
